inventory-service/pyproject.toml
[tool.poetry]
name = "inventory-service"
version = "0.1.0"
description = ""
authors = ["hamzah-syed <hamzah.syed17@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.111.1"
uvicorn = "^0.30.3"
sqlmodel = "^0.0.21"
kafka-python = "^2.0.2"
pydantic = "^2.8.2"
psycopg2-binary = "^2.9.9"
protobuf = "^5.27.2"
dapr = "^1.13.0"
python-dotenv = "^1.0.1"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.2"
httpx = "^0.27.0"
pytest-asyncio = "^0.23.8"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
